modified share-capture
