+++ gradle/libs.versions.toml
@@ -5,7 +5,7 @@ agp = "8.7.2"
 androidDesugarJdkLibs = "2.0.4"
 androidxPaging = "3.1.0"
 androidx-test-ext-junit = "1.1.5"
-capturable = "2.1.0"
+capturable = "3.0.1"
 coil-compose = "2.4.0"
 compose-bom = "2025.03.00"
 constraintlayout-compose = "1.0.1"
